Refactor participant limits and oracle hash handling in coinflip program
- Changed the types of `max_participants` and `min_participants` from `u8` to `u16` across multiple files to accommodate a larger number of participants.
- Updated the handling of `oracle_hash` to store it directly as a `[u8; 32]` instead of an `Option`, simplifying the logic for checking its presence.
- Adjusted related account space calculations to reflect these changes, improving memory efficiency and clarity in game initialization and state management.

These changes enhance the functionality and maintainability of the coinflip program, allowing for more participants and clearer oracle hash management.

+++ programs/coinflip/src/instructions/initialize_game.rs
@@ -7,8 +7,8 @@ pub fn handler(
     ctx: Context<super::InitializeGame>,
     game_type: GameType,
     amount: u64,
-    max_participants: u8,
-    min_participants: u8,
+    max_participants: u16,
+    min_participants: u16,
     timeout_duration: i64,
     is_private: bool,
 ) -> Result<()> {
